+++ examples/examples_test.go
@@ -1,7 +1,23 @@
 package examples
+
 import (
+	"os"
 	"testing"
+
+	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
 )
-func TestExamples(t *testing.T) {
-	t.Skip("Not implemented")
+
+func getCwd(t *testing.T) string {
+	cwd, err := os.Getwd()
+	if err != nil {
+		t.FailNow()
+	}
+
+	return cwd
+}
+
+func getBaseOptions(t *testing.T) integration.ProgramTestOptions {
+	return integration.ProgramTestOptions{
+		ExpectRefreshChanges: true,
+	}
 }