+++ provider/cmd/pulumi-resource-hcloud/main.go
@@ -19,11 +19,11 @@ package main
 import (
 	"github.com/pulumi/pulumi-terraform-bridge/v2/pkg/tfbridge"
 
-	xyz "github.com/pulumi/pulumi-xyz/provider"
-	"github.com/pulumi/pulumi-xyz/provider/pkg/version"
+	hcloud "github.com/pulumi/pulumi-hcloud/provider"
+	"github.com/pulumi/pulumi-hcloud/provider/pkg/version"
 )
 
 func main() {
 	// Modify the path to point to the new provider
-	tfbridge.Main("xyz", version.Version, xyz.Provider(), pulumiSchema)
+	tfbridge.Main("hcloud", version.Version, hcloud.Provider(), pulumiSchema)
 }