+++ provider/resources.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package xyz
+package hcloud
 
 import (
 	"unicode"
@@ -22,13 +22,13 @@ import (
 	"github.com/pulumi/pulumi-terraform-bridge/v2/pkg/tfbridge"
 	"github.com/pulumi/pulumi/sdk/v2/go/common/resource"
 	"github.com/pulumi/pulumi/sdk/v2/go/common/tokens"
-	"github.com/terraform-providers/terraform-provider-xyz/xyz"
+	"github.com/terraform-providers/terraform-provider-hcloud/hcloud"
 )
 
 // all of the token components used below.
 const (
 	// packages:
-	mainPkg = "xyz"
+	mainPkg = "hcloud"
 	// modules:
 	mainMod = "index" // the y module
 )
@@ -59,20 +59,6 @@ func makeResource(mod string, res string) tokens.Type {
 	return makeType(mod+"/"+fn, res)
 }
 
-// boolRef returns a reference to the bool argument.
-func boolRef(b bool) *bool {
-	return &b
-}
-
-// stringValue gets a string value from a property map if present, else ""
-func stringValue(vars resource.PropertyMap, prop resource.PropertyKey) string {
-	val, ok := vars[prop]
-	if ok && val.IsString() {
-		return val.StringValue()
-	}
-	return ""
-}
-
 // preConfigureCallback is called before the providerConfigure function of the underlying provider.
 // It should validate that the provider can be configured, and provide actionable errors in the case
 // it cannot be. Configuration variables can be read from `vars` using the `stringValue` function -
@@ -81,55 +67,49 @@ func preConfigureCallback(vars resource.PropertyMap, c *terraform.ResourceConfig
 	return nil
 }
 
-// managedByPulumi is a default used for some managed resources, in the absence of something more meaningful.
-var managedByPulumi = &tfbridge.DefaultInfo{Value: "Managed by Pulumi"}
-
 // Provider returns additional overlaid schema and metadata associated with the provider..
 func Provider() tfbridge.ProviderInfo {
 	// Instantiate the Terraform provider
-	p := xyz.Provider().(*schema.Provider)
+	p := hcloud.Provider().(*schema.Provider)
 
 	// Create a Pulumi provider mapping
 	prov := tfbridge.ProviderInfo{
-		P:           p,
-		Name:        "xyz",
-		Description: "A Pulumi package for creating and managing xyz cloud resources.",
-		Keywords:    []string{"pulumi", "xyz"},
-		License:     "Apache-2.0",
-		Homepage:    "https://pulumi.io",
-		Repository:  "https://github.com/pulumi/pulumi-xyz",
-		Config:      map[string]*tfbridge.SchemaInfo{
-			// Add any required configuration here, or remove the example below if
-			// no additional points are required.
-			// "region": {
-			// 	Type: makeType("region", "Region"),
-			// 	Default: &tfbridge.DefaultInfo{
-			// 		EnvVars: []string{"AWS_REGION", "AWS_DEFAULT_REGION"},
-			// 	},
-			// },
-		},
+		P:                    p,
+		Name:                 "hcloud",
+		Description:          "A Pulumi package for creating and managing hcloud cloud resources.",
+		Keywords:             []string{"pulumi", "hcloud"},
+		License:              "Apache-2.0",
+		Homepage:             "https://pulumi.io",
+		Repository:           "https://github.com/pulumi/pulumi-hcloud",
+		Config:               map[string]*tfbridge.SchemaInfo{},
 		PreConfigureCallback: preConfigureCallback,
-		Resources:            map[string]*tfbridge.ResourceInfo{
-			// Map each resource in the Terraform provider to a Pulumi type. Two examples
-			// are below - the single line form is the common case. The multi-line form is
-			// needed only if you wish to override types or other default options.
-			//
-			// "aws_iam_role": {Tok: makeResource(mainMod, "IamRole")}
-			//
-			// "aws_acm_certificate": {
-			// 	Tok: makeResource(mainMod, "Certificate"),
-			// 	Fields: map[string]*tfbridge.SchemaInfo{
-			// 		"tags": {Type: makeType(mainPkg, "Tags")},
-			// 	},
-			// },
+		Resources: map[string]*tfbridge.ResourceInfo{
+			"hcloud_server":                 {Tok: makeResource(mainMod, "Server")},
+			"hcloud_server_network":         {Tok: makeResource(mainMod, "ServerNetwork")},
+			"hcloud_floating_ip":            {Tok: makeResource(mainMod, "FloatingIp")},
+			"hcloud_floating_ip_assignment": {Tok: makeResource(mainMod, "FloatingIpAssignment")},
+			"hcloud_rdns":                   {Tok: makeResource(mainMod, "Rdns")},
+			"hcloud_ssh_key":                {Tok: makeResource(mainMod, "SshKey")},
+			"hcloud_volume":                 {Tok: makeResource(mainMod, "Volume")},
+			"hcloud_volume_attachment":      {Tok: makeResource(mainMod, "VolumeAttachment")},
+			"hcloud_network":                {Tok: makeResource(mainMod, "Network")},
+			"hcloud_network_route":          {Tok: makeResource(mainMod, "NetworkRoute")},
+			"hcloud_network_subnet":         {Tok: makeResource(mainMod, "NetworkSubnet")},
 		},
 		DataSources: map[string]*tfbridge.DataSourceInfo{
-			// Map each resource in the Terraform provider to a Pulumi function. An example
-			// is below.
-			// "aws_ami": {Tok: makeDataSource(mainMod, "getAmi")},
+			"hcloud_datacenter":  {Tok: makeDataSource(mainMod, "getDatacenter")},
+			"hcloud_datacenters": {Tok: makeDataSource(mainMod, "getDatacenters")},
+			"hcloud_floating_ip": {Tok: makeDataSource(mainMod, "getFloatingIp")},
+			"hcloud_image":       {Tok: makeDataSource(mainMod, "getImage")},
+			"hcloud_location":    {Tok: makeDataSource(mainMod, "getLocation")},
+			"hcloud_locations":   {Tok: makeDataSource(mainMod, "getLocations")},
+			"hcloud_server":      {Tok: makeDataSource(mainMod, "getServer")},
+			"hcloud_ssh_key":     {Tok: makeDataSource(mainMod, "getSshKey")},
+			"hcloud_network":     {Tok: makeDataSource(mainMod, "getNetwork")},
+			"hcloud_ssh_keys":    {Tok: makeDataSource(mainMod, "getSshKeys")},
+			"hcloud_volume":      {Tok: makeDataSource(mainMod, "getVolume")},
 		},
 		JavaScript: &tfbridge.JavaScriptInfo{
-			AsyncDataSources: true,
 			// List any npm dependencies and their versions
 			Dependencies: map[string]string{
 				"@pulumi/pulumi": "^2.0.0",
@@ -153,6 +133,9 @@ func Provider() tfbridge.ProviderInfo {
 			PackageReferences: map[string]string{
 				"Pulumi":                       "2.*",
 				"System.Collections.Immutable": "1.6.0",
+			},
+			Namespaces: map[string]string{
+				mainPkg: "HCloud",
 			},
 		},
 	}