+++ sdk/go/hcloud/doc.go
@@ -1,0 +1,3 @@
+// A Pulumi package for creating and managing hcloud cloud resources.
+//
+package hcloud