+++ CHANGELOG.md
@@ -2,5 +2,6 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
+* Initial creation of the provider using v1.16.0 of the Hetzner Cloud Terraform Provider
 
 ---